feat: extract stroke/edge-line utils to marbleVisuals.ts and apply to Rush mode
Move attachBallStroke, attachEdgeLines, and addEdgeLinesToScene into a
shared marbleVisuals.ts (with hasEdgeLines guard to prevent double-apply
on repeated traversals). Import in both Race and Rush composables so
Rush mode renders toon outlines on the marble and all spawned platforms.
